+++ CHANGELOG.md
@@ -2,4 +2,5 @@
 This is the changelog for [OpenAPI Factory](readme.md).
 
 ## 1.0 ##
-Created npm library for building apis.
+* Created npm library for building apis.
+* Pull api method and resource from AWS context.